mode all validations to middleware.js
