+++ middleware.js
@@ -1,3 +1,9 @@
+const { campgroundSchema } = require('./schemas.js');
+const { reviewSchema } = require('./schemas.js');
+
+const ExpressError = require('./utils/ExpressError');
+const Campground = require('./models/campground');
+
 module.exports.isLoggedIn = (req, res, next) => {
 
     if (!req.isAuthenticated()) {
@@ -9,3 +15,47 @@ module.exports.isLoggedIn = (req, res, next) => {
 
     next();
 }
+
+module.exports.validateCampground = (req, res, next) => {
+
+    const { error } = campgroundSchema.validate(req.body);
+
+    if (error) {
+
+        const message = error.details.map(el => el.message).join(',');
+        throw new ExpressError(message, 400);
+
+    } else {
+
+        next();
+    }
+};
+
+module.exports.isAuthor = async (req, res, next) => {
+
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+
+    if (!campground.author.equals(req.user._id)) {
+
+        req.flash('error', 'You do not have permission to do that!');
+        res.redirect(`/campgrounds/${id}`);
+    }
+
+    next();
+};
+
+module.exports.validateReview = (req, res, next) => {
+
+    const { error } = reviewSchema.validate(req.body);
+
+    if (error) {
+
+        const message = error.details.map(el => el.message).join(',');
+        throw new ExpressError(message, 400);
+
+    } else {
+
+        next();
+    }
+};
