+++ routes/reviews.js
@@ -5,24 +5,9 @@ const catchAsync = require('../utils/catchAsync');
 const Campground = require('../models/campground');
 const Review = require('../models/review');
 
-const { reviewSchema } = require('../schemas.js');
-const ExpressError = require('../utils/ExpressError');
+const { validateReview } = require('../middleware');
 
 
-const validateReview = (req, res, next) => {
-
-    const { error } = reviewSchema.validate(req.body);
-
-    if (error) {
-
-        const message = error.details.map(el => el.message).join(',');
-        throw new ExpressError(message, 400);
-
-    } else {
-
-        next();
-    }
-};
 
 router.post('/', validateReview, catchAsync(async (req, res) => {
