+++ views/campgrounds/show.ejs
@@ -20,12 +20,14 @@
                 </li>
                 <li class="list-group-item">$<%= campground.price %>/night</li>
             </ul>
+            <% if ( currentUser && campground.author.equals(currentUser._id)) { %>
             <div class="card-body">
                 <a class="card-link btn btn-info" href="/campgrounds/<%= campground._id %>/edit">Edit</a>
                 <form class="d-inline" action="/campgrounds/<%= campground._id %>?_method=DELETE" method="POST">
                     <button class="btn btn-warning">Delete</button>
                 </form>
             </div>
+            <% } %>
         </div>
     </div>
     <div class="col-6">
